ft(Validation): provide a mechanism of injecting validator middleware into routes definitions
refactor app architecture by moving generic modules into the app folder
use express-validator declarative method run(req) in the validate
middleware. This makes it possible to inject validation logic into the
routes definitions by decoupling the validate function from the
validation rules it runs
The above capability allows for rules to be injected into the generic
validate middleware on a per route definition basis and further  allows
the validator middleware itself to be injectable into a route as a dependency.